fix interactors
