+++ .rspec
@@ -1,1 +1,4 @@
+--color
 --require spec_helper
+--profile
+--format documentation
