--- app/interactors/projects/destroy_organize.rb
@@ -1,8 +1,0 @@
-module Projects
-  class Destroy
-    include Interactor::Organizer
-
-    organize Projects::Destroy::Execute,
-             Projects::Destroy::SendNotification
-  end
-end
