+++ app/interactors/projects/update.rb
@@ -1,13 +1,8 @@
 module Projects
   class Update
-    class Execute
-      include Interactor
+    include Interactor::Organizer
 
-      delegate :project, :project_params, to: :context
-
-      def call
-        context.fail!(error: "Invalid Data") unless project.update(project_params)
-      end
-    end
+    organize Projects::Update::Execute,
+             Projects::Update::SendNotification
   end
 end
