--- app/interactors/projects/update_organize.rb
@@ -1,8 +1,0 @@
-module Projects
-  class Update
-    include Interactor::Organizer
-
-    organize Projects::Update::Execute,
-             Projects::Update::SendNotification
-  end
-end
